+++ src/diagnostics/tsx/validators/Image.ts
@@ -2,7 +2,7 @@ import { ALT, GENERIC_ALT, IMG } from '../../utils/constants';
 import { messages } from '../../utils/messages';
 import { Diagnostic } from '../Diagnostic';
 import { TSXElement } from '../Element';
-import { Validator } from './Validator';
+import { Validator, Visitor } from './Validator';
 
 export class ImageValidator implements Validator {
   #tags = [IMG];
@@ -31,5 +31,9 @@ export class ImageValidator implements Validator {
     if (altText && hasGenericAlt) {
       return new Diagnostic(messages.img.generic + altText, node.loc);
     }
+  }
+
+  accept<T>(visitor: Visitor<T>, node: TSXElement): T {
+    return visitor.validateImage(node);
   }
 }